got automatic entity removal working on map
in the future this should possibly be changed to happen on server side, though
that doesn't seem flawless either - maybe it's better this way.

Also, I saw that the explore thing is seriously flawed - we need to make this
code use the same collections that the mapview and other views use.

Anyway, this closes #42.